Add files via upload
We are working on a deep learning project to develop a Class Monitoring System using Convolutional Neural Networks (CNN). The system captures images in the classroom and analyzes them to detect cheating and fighting. This project aims to provide real-time monitoring and alerts to maintain a secure and disciplined classroom environment.